requirements.txt
PyYAML==5.4.1
requests==2.26.0
symbol-sdk-core-python==2.0.1
zenlog==1.1
